Move billing-history top-level components to own file
