+++ client/my-sites/purchases/controller.js
@@ -14,9 +14,8 @@ import {
 	PurchaseCancelDomain,
 	PurchaseAddPaymentMethod,
 	PurchaseEditPaymentMethod,
-	BillingHistory,
-	ReceiptView,
-} from 'my-sites/purchases/main.tsx';
+} from 'my-sites/purchases/main';
+import { BillingHistory, ReceiptView } from 'my-sites/purchases/billing-history';
 
 export function redirectToPurchases( context ) {
 	const siteDomain = context.params.site;
